+++ odoo/local-src/sale_discount_program/__manifest__.py
@@ -20,7 +20,9 @@
         'data/ir_config_parameter.xml',
         'data/product_voucher.xml',
         'data/program_sequence.xml',
+        # Security
         'security/ir.model.access.csv',
+        # Views
         'views/program.xml',
         'views/program_action.xml',
         'views/program_condition.xml',